core: Separate operation.rs into modules
