+++ p2panda-auth/src/traits/resolver.rs
@@ -2,25 +2,24 @@
 
 use std::fmt::Debug;
 
-use p2panda_core::identity::Author;
-use p2panda_core::traits::OperationId;
+use p2panda_core::traits::{Author, OperationId};
 
-/// Interface for implementing a custom group crdt resolver.
+/// Interface for implementing a custom group CRDT resolver.
 pub trait Resolver<ID, OP, M, C>
 where
     ID: Author,
-    OP: OperationId + Ord,
+    OP: OperationId,
 {
     type State;
+
     type Error: Debug;
 
-    /// Check if this message requires that a full state re-build takes place. This would usually
-    /// be due to concurrent operations arriving which require special handling.
+    /// Check if this message requires that a full state re-build takes place. This would usually be
+    /// due to concurrent operations arriving which require special handling.
     fn rebuild_required(y: &Self::State, msg: &M) -> Result<bool, Self::Error>;
 
     /// Process all operations and update internal state as required.
     ///
     /// This could include updating any internal filter object.
-    #[allow(clippy::type_complexity)]
     fn process(y: Self::State) -> Result<Self::State, Self::Error>;
 }
